# target/classes/swaggerapi.yml
swagger: '2.0'
info:
  title: SpringBoot Test API
  description: |
    Spring boot API services for learning
  version: 0.01.5
host: localhost
schemes:
  - https
basePath: /api/services
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /cities/getAllCities:
    get:
      summary: Get All cities
      deprecated: false
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/Cities'
        401:
          description: Not authenticated. All activities are logged per user, so users must be authenticated before they can call this endpoint.
        403:
          description: Not authorized for any data. Note this is not the case if user has a role of Administrator or Manager.
        404:
          description: The origin server did not find a current representation for the target resource or is not willing to disclose that one exists.
        405:
          description: The method received in the request-line is known by the origin server but not supported by the target resource.
        500:
          description: The server encountered an unexpected condition that prevented it from fulfilling the request.
        503:
          description: The server is currently unable to handle the request due to a temporary overload or scheduled maintenance, which will likely be alleviated after some delay
        default:
          description: Unexpected ErrorResponse
definitions:
...
  Cities:
    type: array
    title: All cities details
    items:
      $ref: '#/definitions/City'
  City:
    type: object
    properties:
      cityId:
        type: integer
        description: city Id
      cityName:
        type: string
        description: city name